refactor(zlog): improve code quality and maintainability
- Remove duplicate code
- Improve error handling
- Optimize imports
- Enhance readability and formatting
